Refactor of stochastic engines, including serial engines
