+++ ptypy/engines/SDR.py
@@ -12,7 +12,6 @@ import time
 from .. import utils as u
 from ..utils.verbose import logger, log
 from ..utils import parallel
-from .utils import basic_fourier_update
 from . import register
 from .stochastic import StochasticBaseEngine
 from ..core.manager import Full, Vanilla, Bragg3dModel, BlockVanilla, BlockFull
@@ -66,6 +65,14 @@ class SDR(StochasticBaseEngine):
         """
         super().__init__(ptycho_parent, pars)
 
+        # SDR Adjustment parameters
+        self.alpha = self.p.sigma # TODO: replace with generic fourier update params
+        self.tau = self.p.tau # TODO: replace with generic fourier update params
+        self.prA = self.p.beta_probe
+        self.prB = 0.0
+        self.obA = self.p.beta_object
+        self.obB = 0.0
+
         self.ptycho.citations.add_article(
             title='Semi-implicit relaxed Douglas-Rachford algorithm (sDR) for ptychography',
             author='Pham et al.',
@@ -76,35 +83,3 @@ class SDR(StochasticBaseEngine):
             doi='10.1364/OE.27.031246',
             comment='The semi-implicit relaxed Douglas-Rachford reconstruction algorithm',
         )
-
-    def fourier_update(self, view):
-        """
-        Fourier update for Stochastic Douglas-Rachford (SDR).
-
-
-        """
-        return basic_fourier_update(view, alpha=self.p.sigma, tau=self.p.tau, 
-                                    LL_error=self.p.compute_log_likelihood)
-
-    def object_update(self, *args, **kwargs):
-        """
-        Object update for SDR.
-
-        .. math::
-            O^{j+1} += \\alpha * \\bar{P^{j}} * (\\Psi^{\prime} - \\Psi^{j}) / P_{norm}
-            P_{norm} = (1 - \\alpha) * ||P^{j}||_{max}^2 + \\alpha * |P^{j}|^2
-
-        """
-        self.generic_object_update(*args, **kwargs, A=self.p.beta_object, B=0.0)
-
-
-    def probe_update(self, *args, **kwargs):
-        """
-        Probe update for SDR. 
-
-        .. math::
-            P^{j+1} += \\alpha * \\bar{O^{j}} * (\\Psi^{\prime} - \\Psi^{j}) / O_{norm}
-            O_{norm} = (1 - \\alpha) * ||O^{j}||_{max}^2 + \\alpha * |O^{j}|^2
-
-        """
-        self.generic_probe_update(*args, **kwargs, A=self.p.beta_probe, B=0.0)
